requirements.txt
Django
djangorestframework
requests
django-redis
django-environ
icecream
drf-spectacular
drf-spectacular-sidecar